v1.1.4 - Fix cleaning bug

+++ cctyper/controller.py
@@ -45,7 +45,7 @@ class Controller(object):
             logging.basicConfig(format='[%(asctime)s] %(levelname)s: %(message)s', datefmt='%Y-%m-%d %H:%M:%S', level=self.lvl)
         else:
             logging.basicConfig(format='\033[36m'+'[%(asctime)s] %(levelname)s:'+'\033[0m'+' %(message)s', datefmt='%Y-%m-%d %H:%M:%S', level=self.lvl)
-        logging.info('Running CRISPRCasTyper version 1.1.3')
+        logging.info('Running CRISPRCasTyper version 1.1.4')
 
         # Force consistency
         self.out = os.path.join(self.out, '')
@@ -118,9 +118,10 @@ class Controller(object):
             if os.stat(self.out+'hmmer.log').st_size == 0:
                 os.remove(self.out+'hmmer.log')
 
-            if os.stat(self.out+'hmmer_custom.log').st_size == 0:
-                os.remove(self.out+'hmmer_custom.log')
-            
+            if self.customhmm != '':
+                if os.stat(self.out+'hmmer_custom.log').st_size == 0:
+                    os.remove(self.out+'hmmer_custom.log')
+                
             if not self.keep_tmp:
                 
                 logging.info('Removing temporary files')
